refactor: combine two types of question form
